integrate iwmfdiff
