+++ functions/ddrm/configs/imagenet_512_cc.yml
