+++ functions/ddrm/inp_masks/lolcat_extra.npy
