+++ functions/ddrm/inp_masks/lorem3.npy
